feat: impl backend para UH05 e uso de double para anos

+++ src/main/java/br/com/edengreen_impact/model/CartaoInput.java
@@ -1,3 +1,3 @@
 package br.com.edengreen_impact.model;
 
-public record CartaoInput(String tipo, int anos) {}
+public record CartaoInput(String tipo, double anos) {}
